refactor(frontend): update persona routing and consolidate delete functionality
