+++ chatbot-core/frontend/src/app/assistants/edit/[id]/page.tsx
@@ -5,7 +5,7 @@ import { HeaderWrapper } from "@/components/header/HeaderWrapper";
 import { AssistantEditor } from "@/app/admin/assistants/AssistantEditor";
 import { SuccessfulPersonaUpdateRedirectType } from "@/app/admin/assistants/enums";
 import { fetchAssistantEditorInfoSS } from "@/lib/assistants/fetchPersonaEditorInfoSS";
-import { DeletePersonaButton } from "@/app/admin/assistants/[id]/DeletePersonaButton";
+import { DeletePersonaButton } from "@/app/admin/assistants/DeletePersonaButton";
 import { LargeBackButton } from "../../LargeBackButton";
 import Title from "@/components/ui/title";
